// src/library/components/molecules/multi-combo-box/types.ts
export interface InputOption<ValueType> {
  label?: string;
  value: ValueType;
  disabled?: boolean;
  [key: string]: any;
}

export interface InternalInputOption<ValueType> extends InputOption<ValueType> {
  label: string;
  id: string;
}

export interface State<ValueType> {
  isOpen: boolean;
  filteredOptions: InternalInputOption<ValueType>[];
  options: InternalInputOption<ValueType>[];
  selectedOptions: InternalInputOption<ValueType>[];
  inputValue: string;
}

export type StateActions<ValueType> =
  | { type: "open" }
  | { type: "close" }
  | { type: "select"; id: string }
  | { type: "deSelect"; id: string }
  | { type: "filter"; payload: InternalInputOption<ValueType>[] }
  | { type: "clear" }
  | { type: "onExternalValueChange"; value: ValueType[] | undefined }
  | { type: "onInputChange"; value: string }
  | { type: "onBackspace" };